fix warnings

+++ ColorChecker_imaged_values.m
@@ -9,8 +9,8 @@ ys = linspace(245, 1675, 4) - (width / 2);
 
 [x, y] = meshgrid(ys, xs); % row col, x y
 
-rx = reshape(x, [24, 1]);
-ry = reshape(y, [24, 1]);
+rx = uint16(reshape(x, [24, 1]));
+ry = uint16(reshape(y, [24, 1]));
 
 means = zeros([24 3]);
 for idx = 1:24
